Implemented Manage Alerts UI, Backend to active or deactivate a alert

+++ src/main/java/com/cricFizzApp/bean/alert/AlertDetails.java
@@ -37,7 +37,7 @@ public class AlertDetails {
 
     @CassandraType(type = CassandraType.Name.INT)
     @Column("seriesId")
-    private Long seriesId;
+    private Integer seriesId;
 
     @CassandraType(type = CassandraType.Name.TEXT)
     @Column("seriesName")
@@ -45,7 +45,7 @@ public class AlertDetails {
 
     @CassandraType(type = CassandraType.Name.INT)
     @Column("matchId")
-    private Long matchId;
+    private Integer matchId;
 
     @CassandraType(type = CassandraType.Name.TEXT)
     @Column("matchName")
